% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{app_secret_ask}
\alias{app_secret_ask}
\title{app_secret_ask}
\usage{
app_secret_ask(obj = NULL, file = NULL,
  message = "Please provide your password")
}
\arguments{
\item{obj}{An AppSecret object}

\item{file}{path to the file to write encrypted data to}

\item{message}{Message to present the user with}
}
\value{
Boolean as to whether user was asked for input.
}
\description{
Utility function designed for use as a standalone, or in \code{tap()}
}
\examples{

\dontrun{
  asm       <- app_secret_manager()
  pass_file <- asm$path_in_vault("password")
  password  <- asm$tap(function(obj, file, message) {
     if(app_secret_ask(asm = obj, file = file, message = message)) {
        message("password stored in ", file)
     }
  }, file = pass_file)$decrypt_file(pass_file)

  ## same, but without a message
  password <-
    asm$tap(func = app_secret_ask,
            file = pass_file)$decrypt_file(pass_file)

  ## or just as a save the password helper
  app_secret_ask(obj = asm, file = pass_file)
}

}
