% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_secret.R
\name{app_secret_manager}
\alias{app_secret_manager}
\title{app_secret_manager}
\usage{
app_secret_manager(symmetric_file, key_file)
}
\arguments{
\item{symmetric_file}{path to symmetric file}

\item{key_file}{path to PEM format key file}

\item{...}{compatibility with \code{app_secret_paths} and \code{do.call}}
}
\value{
app_secret instance
}
\description{
app_secret_manager
}
