% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{app_secret_paths}
\alias{app_secret_paths}
\title{app_secret_paths}
\usage{
app_secret_paths(appname = NULL, base_path = Sys.getenv("HOME"))
}
\arguments{
\item{appname}{The application name}

\item{base_path}{The base path to use for storing the symmetric and encrypted files}
}
\value{
list()

Influential environment variables \code{APP_SECRET_USE_HERE}
}
\description{
Create paths as a utility and set of rules/expectations. Useful when trying to enforce a convention.
}
\examples{

\dontrun{
  # list(symmetric_file = "/home/user/.best-shiny-app-ever/symmetric.rsa",
  #      key_file       = "/home/user/.best-shiny-app-ever/secret.pem",
  #      vault          = "/home/user/.best-shiny-app-ever")

  Sys.setenv("APP_SECRET_USE_HERE"="1")
  here::set_here(path = "/apps/another-shiny-app")
  # list(symmetric_file = "/apps/another-shiny-app/.user-settings/user.symmetric.rsa",
  #      key_file       = "/home/user/.best-shiny-app-ever/secret.pem",
  #      vault          = "/apps/another-shiny-app/.user-settings")
  paths <- app_secret_paths(appname = "another-shiny-app")

  Sys.unsetenv("APP_SECRET_USE_HERE")
  here::set_here(path = "/apps/dull-app")
  # list(symmetric_file = "/apps/dull-app/.dull-app/user.symmetric.rsa",
  #      key_file       = "/home/user/.dull-app/secret.pem",
  #      vault          = "/apps/dull-app/.dull-app")
  paths <- app_secret_paths(appname = "dull-app", base_path = file.path(here()))
}

}
